inphora-fixes/inphora-fixes/backend/requirements.txt
# Inphora Backend — requirements.txt
# FIXED: Python 3.12 compatible, pinned versions for reproducibility

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.8          # Required by pymysql for SSL connections

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Rate limiting — CRIT-05
slowapi==0.1.9

# Validation
pydantic[email]==2.7.1
pydantic-settings==2.3.1

# Structured logging
structlog==24.2.0

# HTTP client (for webhooks etc.)
httpx==0.27.0

# CORS & middleware
python-multipart==0.0.9        # Required for form data (file uploads)

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0                  # Used by TestClient
